#1334 cleanup button
