+++ lib/elementWrapper/buttonWrapper.js
@@ -1,18 +1,19 @@
-const CheckBox = require('../elements/checkBox');
-const { $function } = require('../elementSearch');
+const { $function, match } = require('../elementSearch');
 const ElementWrapper = require('./elementWrapper');
-const { firstElement, getElementGetter } = require('./helper');
+const { getElementGetter } = require('./helper');
 
-function getCheckBoxElementWithLabel(label) {
-  const checkBoxes = [];
+function getButtonElementWithLabel(label) {
+  const buttons = [];
+  const inputTypes = ['button', 'submit', 'reset', 'image'];
   function checkAndPushElement(elem) {
     if (
-      elem.tagName &&
-      elem.tagName.toLowerCase() == 'input' &&
-      elem.type &&
-      elem.type.toLowerCase() === 'checkbox'
+      (elem.tagName && elem.tagName.toLowerCase() === 'button') ||
+      (elem.tagName &&
+        elem.tagName.toLowerCase() == 'input' &&
+        elem.type &&
+        inputTypes.includes(elem.type.toLowerCase()))
     ) {
-      checkBoxes.push(elem);
+      buttons.push(elem);
     }
   }
   const matchingLabels = [...document.querySelectorAll('label')].filter((labelElem) => {
@@ -31,46 +32,37 @@ function getCheckBoxElementWithLabel(label) {
       });
     }
   }
-  //check label is inlined
-  const inputRadioElements = [...document.querySelectorAll('input[type="checkbox" i]')];
-  for (let inputRadioElement of inputRadioElements) {
-    if (
-      inputRadioElement.nextSibling.nodeType === Node.TEXT_NODE &&
-      inputRadioElement.nextSibling.wholeText.includes(label)
-    ) {
-      checkBoxes.push(inputRadioElement);
-    }
-  }
-  return checkBoxes;
+  // check inputs matching types and value
+  const matchingValues = document.querySelectorAll(
+    `input[type="submit"][value*="${label}"],input[type="image"][value*="${label}"],input[type="reset"][value*="${label}"],input[type="button"][value*="${label}"]`,
+  );
+  return buttons.concat([...matchingValues]);
 }
-class CheckBoxWrapper extends ElementWrapper {
-  _get = getElementGetter(
+class ButtonWrapper extends ElementWrapper {
+  getByButton = getElementGetter(
     this.selector,
-    async () =>
-      await $function(
-        getCheckBoxElementWithLabel,
-        this.selector.label,
-        this.options.selectHiddenElements,
-      ),
-    'input[type="checkbox" i]',
+    async () => match(this.selector.label, this.options).elements('button', 0, 0),
+    'button',
     this.options.selectHiddenElements,
   );
 
-  async isChecked() {
-    const elem = await firstElement.apply(this);
-    return await elem.isChecked();
-  }
-  async check() {
-    const elem = await firstElement.apply(this);
-    await elem.check();
-  }
-  async uncheck() {
-    const elem = await firstElement.apply(this);
-    await elem.uncheck();
-  }
-  async elements(retryInterval, retryTimeout) {
-    let elements = await super.elements(retryInterval, retryTimeout);
-    return elements.map((element) => CheckBox.from(element, this._description));
-  }
+  getByInput = getElementGetter(
+    this.selector,
+    async () =>
+      await $function(
+        getButtonElementWithLabel,
+        this.selector.label,
+        this.options.selectHiddenElements,
+      ),
+    'input[type="submit"],input[type="reset"],input[type="button"],input[type="image"]',
+    this.options.selectHiddenElements,
+  );
+  _get = async () => {
+    const input = await this.getByInput();
+    if (input.length) {
+      return input;
+    }
+    return this.getByButton();
+  };
 }
-module.exports = CheckBoxWrapper;
+module.exports = ButtonWrapper;
