+++ lib/elementWrapper/elementWrapperFactory.js
@@ -1,5 +1,6 @@
 const ElementWrapper = require('./elementWrapper');
 const FileFieldWrapper = require('./fileFieldWrapper');
+const ButtonWrapper = require('./buttonWrapper');
 const RadioButtonWrapper = require('./radioButtonWrapper');
 const TextBoxWrapper = require('./textBoxWrapper');
 const CheckBoxWrapper = require('./checkBoxWrapper');
@@ -15,6 +16,8 @@ const createElementWrapper = (elementType, query, attrValuePairs, _options, ...a
   switch (elementType) {
     case 'FileField':
       return new FileFieldWrapper(selector, options, description);
+    case 'Button':
+      return new ButtonWrapper(selector, options, description);
     // case 'textBox':
     //   return new TextBoxWrapper(get, description, getIfExists);
     // case 'radioButton':
