+++ lib/taiko.js
@@ -2175,82 +2175,7 @@ module.exports.listItem = (attrValuePairs, _options = {}, ...args) => {
  */
 module.exports.button = (attrValuePairs, _options = {}, ...args) => {
   validate();
-  let get;
-  const { selector, options } = prepareParameters(attrValuePairs, _options, ...args);
-
-  function getButtonElementWithLabel(label) {
-    const buttons = [];
-    const inputTypes = ['button', 'submit', 'reset', 'image'];
-    function checkAndPushElement(elem) {
-      if (
-        (elem.tagName && elem.tagName.toLowerCase() === 'button') ||
-        (elem.tagName &&
-          elem.tagName.toLowerCase() == 'input' &&
-          elem.type &&
-          inputTypes.includes(elem.type.toLowerCase()))
-      ) {
-        buttons.push(elem);
-      }
-    }
-    const matchingLabels = [...document.querySelectorAll('label')].filter((labelElem) => {
-      return labelElem.innerText.toLowerCase().includes(label.toLowerCase());
-    });
-    for (let matchingLabel of matchingLabels) {
-      const labelFor = matchingLabel.getAttribute('for');
-      if (labelFor) {
-        //check label with attribute for
-        const labelForElement = document.getElementById(labelFor);
-        checkAndPushElement(labelForElement);
-      } else {
-        // check child node of label tag
-        matchingLabel.childNodes.forEach((elem) => {
-          checkAndPushElement(elem);
-        });
-      }
-    }
-    // check inputs matching types and value
-    const matchingValues = document.querySelectorAll(
-      `input[type="submit"][value*="${label}"],input[type="image"][value*="${label}"],input[type="reset"][value*="${label}"],input[type="button"][value*="${label}"]`,
-    );
-    return buttons.concat([...matchingValues]);
-  }
-
-  if (!selector.attrValuePairs && !selector.label) {
-    get = async () =>
-      await handleRelativeSearch(
-        await $$(
-          'input[type="submit"],input[type="reset"],input[type="button"],input[type="image"],button',
-          options.selectHiddenElements,
-        ),
-        selector.args,
-      );
-  } else {
-    const getByButton = getElementGetter(
-      selector,
-      async () => match(selector.label, options).elements('button', 0, 0),
-      'button',
-      options.selectHiddenElements,
-    );
-
-    const getByInput = getElementGetter(
-      selector,
-      async () =>
-        await $function(getButtonElementWithLabel, selector.label, options.selectHiddenElements),
-      'input[type="submit"],input[type="reset"],input[type="button"],input[type="image"]',
-      options.selectHiddenElements,
-    );
-    get = async () => {
-      const input = await getByInput();
-      if (input.length) {
-        return input;
-      }
-      return getByButton();
-    };
-  }
-
-  const description = desc(selector, 'label', 'Button');
-
-  return createElementWrapper(get, description);
+  return createElementWrapper('Button', 'label', attrValuePairs, _options, ...args);
 };
 
 /**
